requirements.txt
flask
flask-cors
flask-socketio
eventlet
requests
python-dotenv
openai